llm-insight-service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
paho-mqtt==1.6.1
pydantic==1.10.14
google-generativeai==0.5.2
python-dateutil==2.8.2
